다시시작

+++ app/src/main/java/com/example/uohih/joowon/database/ExcelData.kt
@@ -1,5 +1,5 @@
 package com.example.uohih.joowon.database
 
-class WorkerData(val no: Int?, val name: String, val joinDate: Int, val phone: String, val use: Int, val total: Int, val picture: String)
+class WorkerData(val no: Int?, val name: String, val joinDate: Int, val phone: String, val use: Double, val total: Double, val picture: String)
 
-class VacationData(val no: Int?, val name: String, val phone: String, val date: Int, val content: String, val use: Int, val total: Int)
+class VacationData(val no: Int?, val name: String, val phone: String, val date: Int, val content: String, val use: Double, val total: Double)
